requirements.txt
pandas
numpy
matplotlib
seaborn
google-generativeai
python-dotenv
flask
scikit-learn
ipykernel
blinker
click
colorama
Flask-MySQLdb
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
mysqlclient
protobuf
setuptools
Werkzeug
WTForms
blinker
click
colorama
Flask
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
werkzeug
zipp
numpy
tensorflow
Pillow
imbalanced-learn